+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.6.31
+* remove obsolete nodes
+* Set the UI culture back to default after dynamo is closed
+
 ## 0.6.30
 * Update Dynamo Core to 3.4.0-beta6752
 * Move D4R content to en-US folders